web_api/requirements.txt
alembic==1.0.8
aniso8601==6.0.0
astroid==2.2.5
certifi==2019.3.9
chardet==3.0.4
Click==7.0
Flask==1.0.2
Flask-JWT-Extended==3.18.0
Flask-Migrate==2.4.0
Flask-RESTful==0.3.7
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-Testing==0.7.1
gunicorn==19.9.0
idna==2.8
isort==4.3.16
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.3.1
Mako==1.0.8
MarkupSafe==1.1.1
mccabe==0.6.1
passlib==1.7.1
psycopg2-binary==2.7.7
PyJWT==1.7.1
pylint==2.3.1
python-dateutil==2.8.0
python-dotenv==0.10.1
python-editor==1.0.4
pytz==2018.9
requests==2.21.0
six==1.12.0
SQLAlchemy==1.3.1
typed-ast==1.3.1
urllib3==1.24.2
Werkzeug==0.15.1
wrapt==1.11.1